include order tweaking to get rid of redefinition warnings. Ensure to
2005-06-17  Hans Breuer  <hans@breuer.org>

	* plug-ins/python/pydia-*.c : include order tweaking to
	get rid of redefinition warnings. Ensure to include <Python.h>
	before any 'system' header, some dragged in by glib.h
	* plug-ins/python/pydia-diagramdata.c(PyDiaDiagramData_Str)
	  pydia-properties.c(PyDiaProperties_Str)
	  plug-ins/python/pydia-property.c(PyDiaProperty_Str) :
	use %p to format pointers
	* plug-ins/python/pydia-geometry.c(PyDiaArrow_Str) :
	dont cast to float for %d
	* plug-ins/python/pydia-handle.c(PyDiaHandle_Connect) :
	use correct pointer type
	* plug-ins/python/pydia-image.c(PyDiaImage_Str) :
	dont discard qualifier aka dont free const strings
	* plug-ins/python/pydia-properties.c : casts to PyCFunction
	* plug-ins/python/pydia-property.c : removed unused function,
	make ensure_quarks() static
	* plug-ins/python/pydia-render.c : removed unused vars, use
	the correct signature for begin_render
	(dia_py_renderer_finalize): pluged a leak
	=> With -fno-strict-aliasing there is no warning left in all of PyDia

+++ plug-ins/python/pydia-geometry.c
@@ -19,8 +19,6 @@
 
 #include <config.h>
 
-#include <glib.h>
-
 #include "pydia-object.h"
 #include "pydia-geometry.h"
 
@@ -343,7 +341,7 @@ PyDiaArrow_Str(PyDiaArrow *self)
     gchar* s = g_strdup_printf ("(%f,%f, %d)",
                                 (float)(self->arrow.width), 
                                 (float)(self->arrow.length),
-                                (float)(self->arrow.type));
+                                (int)(self->arrow.type));
     py_s = PyString_FromString(s);
     g_free(s);
     return py_s;
